Fix to create load and store pipelines together
A Pair stage, an object used in load and store pipeline,
can not be created because load and store pipelines are instantiated
separately. To create these pipelines together, the pair stage can
be created. Use class PairStage to create the pair stage, which has
methods to process load and store events and has a factory method
to create load and store stages, used in PipelineInitializer.