Split Stage into LoadStage and StoreStage
The interface Stage is not used in both load and store pipeline.
So, dedicated stages are prepared for load and store. With the
introduction of these stages, Pipelie and StageContext are also
splitted into dedicated ones. And PairStage is removed. Use
custom class which implements interface LoadStage and StoreStage
instead.